Adjust limit request
