+++ app/src/main/java/com/example/myrecipebook/rest/message/RecipeMessage.kt
@@ -1,23 +1,27 @@
 package com.example.myrecipebook.rest.message
 
-/**
- * Raw network representation of a Recipe as returned by DummyJSON.
- */
- data class RecipeMessage(
-     val id: Int,
-     val name: String,
-     val ingredients: List<String>,
-     val instructions: List<String>,
-     val prepTimeMinutes: Int,
-     val cookTimeMinutes: Int,
-     val servings: Int,
-     val difficulty: String,
-     val cuisine: String,
-     val caloriesPerServing: Int,
-     val tags: List<String>,
-     val userId: Int,
-     val image: String,
-     val rating: Double,
-     val reviewCount: Int,
-     val mealType: List<String>
- )
+data class RecipeMessage(
+    val id: Int,
+    val name: String,
+    val ingredients: List<String>,
+    val instructions: List<String>,
+    val prepTimeMinutes: Int,
+    val cookTimeMinutes: Int,
+    val servings: Int,
+    val difficulty: String,
+    val cuisine: String,
+    val caloriesPerServing: Int,
+    val tags: List<String>,
+    val userId: Int,
+    val image: String,
+    val rating: Double,
+    val reviewCount: Int,
+    val mealType: List<String>
+)
+
+data class RecipesPageMessage(
+    val recipes: List<RecipeMessage>,
+    val total: Int,
+    val skip: Int,
+    val limit: Int
+)
