--- app/src/main/java/com/example/myrecipebook/rest/message/RecipesPageMessage.kt
@@ -1,11 +1,0 @@
-package com.example.myrecipebook.rest.message
-
-/**
- * Wrapper returned by GET /recipes
- */
- data class RecipesPageMessage(
-    val recipes: List<RecipeMessage>,
-    val total: Int,
-    val skip: Int,
-    val limit: Int
- )
